removed secret; revisions for remote T4 manager; revised classifier weights for cpu-only machine
