--- cand_cutter/kill_cand_cutter
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-kill -9 -"$(pgrep run_cand_cutter)"
-
