--- cand_cutter/run_cand_cutter
@@ -1,15 +1,0 @@
-#!/bin/bash
-#rt="false"
-
-#run this script to start background candidate post-processor
-#> $PWD-logfiles/candcutter_log.txt #/home/ubuntu/proj/dsa110-shell/dsa110-nsfrb/process_server/process_log.txt 
-#dask worker tcp://127.0.0.1:$DASKPORT --nworkers 1 --nthreads 1 --name cand_cutter_WRKR > $PWD/.cand_cutter_WRKR_log.txt 2>&1 &
-
-#python run.py --cutout --cluster --classify --toslack --etcd "$@"  #normal operation
-#if [ "$rt" = "true" ]; then
-#	echo "realtime candcutter active"
-#	nice -15 python run.py --realtime --etcd --subimgpix 301 --SNRthresh 6 --useTOA --psfpercentile 70 --maxcands 10  --cluster --avgcluster --psfcluster --classify3D --clusteriters 2 --maxcand --model_weights3D "/dataz/dsa110/nsfrb/dsa110-nsfrb-training/NN_train/enhanced3dcnn_weights_final3.pth" --classcut --toslack --daskaddress "tcp://10.42.0.228:8786" "$@" #--cluster --psfcluster --classify3D
-#else
-echo "offline candcutter active"
-python run.py --etcd --subimgpix 301 --SNRthresh 6 --useTOA --psfpercentile 70 --maxcands 100  --cluster --avgcluster --psfcluster --classify3D --clusteriters 2 --toslack --maxcand --model_weights3D "/dataz/dsa110/nsfrb/dsa110-nsfrb-training/NN_train/enhanced3dcnn_weights_final3.pth" --classcut "$@" #--cluster --psfcluster --classify3D
-#fi
