--- cand_cutter/tmp.txt.bak
@@ -1,5 +1,0 @@
-1
-2
-3
-4
-5
