+++ casa310nsfrb_env.yml
@@ -37,13 +37,14 @@ dependencies:
   - brotli-python=1.0.9=py310h6a678d5_8
   - bzip2=1.0.8=h5eee18b_6
   - c-ares=1.19.1=h5eee18b_0
-  - ca-certificates=2024.9.24=h06a4308_0
+  - ca-certificates=2025.1.31=hbcca054_0
   - cairo=1.16.0=h35add3b_1015
-  - certifi=2024.8.30=py310h06a4308_0
+  - certifi=2025.1.31=pyhd8ed1ab_0
   - charset-normalizer=2.0.4=pyhd3eb1b0_0
   - click=8.1.7=py310h06a4308_0
   - cloudpickle=3.0.0=py310h06a4308_0
   - comm=0.2.1=py310h06a4308_0
+  - conda-pack=0.8.1=pyhd8ed1ab_1
   - contourpy=1.2.0=py310hdb19cb5_0
   - cuda-cudart=12.1.105=0
   - cuda-cupti=12.1.105=0
@@ -57,13 +58,10 @@ dependencies:
   - cudnn=9.3.0.75=h93bb076_0
   - cycler=0.11.0=pyhd3eb1b0_0
   - cytoolz=0.12.2=py310h5eee18b_0
-  - dask=2024.5.0=py310h06a4308_0
-  - dask-core=2024.5.0=py310h06a4308_0
   - dask-expr=1.1.0=py310h06a4308_0
   - dav1d=1.2.1=h5eee18b_0
   - debugpy=1.6.7=py310h6a678d5_0
   - decorator=5.1.1=pyhd3eb1b0_0
-  - distributed=2024.5.0=py310h06a4308_0
   - executing=0.8.3=pyhd3eb1b0_0
   - ffmpeg=6.0.0=gpl_hdbbbd96_103
   - filelock=3.13.1=py310h06a4308_0
@@ -189,7 +187,7 @@ dependencies:
   - numpy-base=1.24.3=py310h8e6c178_0
   - openh264=2.3.1=hcb278e6_2
   - openjpeg=2.5.2=he7f1fd0_0
-  - openssl=3.0.15=h5eee18b_0
+  - openssl=3.4.1=h7b32b05_0
   - orc=1.8.2=hfdbbad2_2
   - p11-kit=0.24.1=hc5aa10d_0
   - packaging=23.2=py310h06a4308_0
@@ -230,7 +228,7 @@ dependencies:
   - s2n=1.3.37=h3358134_0
   - scipy=1.10.1=py310hd5efca6_0
   - seaborn=0.13.2=py310h06a4308_0
-  - setuptools=75.6.0=pyhff2d567_1
+  - setuptools=75.8.0=pyhff2d567_0
   - six=1.16.0=pyhd3eb1b0_1
   - snappy=1.1.10=h6a678d5_1
   - sortedcontainers=2.4.0=pyhd3eb1b0_0
@@ -289,9 +287,12 @@ dependencies:
       - astropy-healpix==1.0.3
       - async-lru==2.0.4
       - async-timeout==4.0.3
+      - asyncio==3.4.3
       - attrs==24.2.0
       - babel==2.16.0
       - beautifulsoup4==4.12.3
+      - bitarray==3.3.1
+      - bitstring==4.3.1
       - bleach==6.1.0
       - build==1.2.2.post1
       - casadata==2024.6.24
@@ -300,11 +301,14 @@ dependencies:
       - cffi==1.17.0
       - codecov==2.1.13
       - configparser==7.1.0
+      - corner==2.2.3
       - coverage==7.5.4
       - cupy-cuda12x==13.2.0
       - cython==0.29.37
+      - dask==2023.5.0
       - defusedxml==0.7.1
       - dill==0.3.9
+      - distributed==2023.5.0
       - docstring-parser==0.16
       - docutils==0.21.2
       - dodgy==0.2.1
@@ -312,12 +316,14 @@ dependencies:
       - dsa110-calib==3.0.1+4.gb9cf416.dirty
       - dsa110-event==0.7.0
       - dsa110-meridian-fs==1.6.5
-      - dsa110-nsfrb==0.1.0+143.g4e89889.dirty
+      - dsa110-nsfrb==0.1.1+4.g932cd8d.dirty
       - dsa110-pyutils==2.1.1
+      - emcee==3.1.6
       - etcd3==0.12.0
       - exceptiongroup==1.2.1
       - fastjsonschema==2.20.0
       - fastrlock==0.8.2
+      - fitsio==0.9.12
       - flake8==7.1.2
       - flake8-polyfill==1.0.2
       - fqdn==1.5.1
@@ -342,6 +348,7 @@ dependencies:
       - jax-cuda12-plugin==0.4.25
       - jaxlib==0.4.25
       - joblib==1.4.2
+      - jplephem==2.22
       - json5==0.10.0
       - jsonpointer==3.0.0
       - jsonschema==4.23.0
@@ -383,6 +390,7 @@ dependencies:
       - opt-einsum==3.3.0
       - overrides==7.7.0
       - pandocfilters==1.5.1
+      - pdat==0.2.2
       - pep8-naming==0.10.0
       - platformdirs==4.2.2
       - pluggy==1.5.0
@@ -408,7 +416,7 @@ dependencies:
       - pytest==8.2.2
       - python-json-logger==2.0.7
       - pytorchdedispersion==0.1.0
-      - pyuvdata==3.0.0
+      - pyuvdata==3.2.1
       - pyzmq==26.1.1
       - referencing==0.35.1
       - requirements-detector==1.3.2
@@ -450,6 +458,7 @@ dependencies:
       - webcolors==24.8.0
       - webencodings==0.5.1
       - websocket-client==1.8.0
+      - websockets==15.0.1
       - yarl==1.9.4
       - your==0.6.7
 prefix: /home/ubuntu/msherman_nsfrb/miniconda/envs/casa310nsfrb
