+++ legacy/cand_cutter_old/kill_cand_cutter
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+kill -9 -"$(pgrep run_cand_cutter)"
+
