+++ nsfrb/config.py
@@ -194,4 +194,5 @@ NSFRB_CANDDADA_BYTES = 144961600
 NSFRB_SRCHDADA_BYTES = 28992320
 NSFRB_TOADADA_BYTES = 28992320
 
-
+minDM = 171
+maxDM = 4000
