+++ nsfrb/outputlogging.py
@@ -10,8 +10,7 @@ cwd = os.environ['NSFRBDIR']
 
 output_file = cwd + "-logfiles/run_log.txt"#"/home/ubuntu/proj/dsa110-shell/dsa110-nsfrb/tmpoutput/run_log.txt"
 """
-from nsfrb.config import cwd,cand_dir,frame_dir,psf_dir,img_dir,vis_dir,raw_cand_dir,backup_cand_dir,final_cand_dir,inject_dir,training_dir,noise_dir,imgpath,coordfile,output_file,processfile,timelogfile,cutterfile,pipestatusfile,searchflagsfile,run_file,processfile,cutterfile,cuttertaskfile,flagfile,error_file,inject_file,recover_file,binary_file
-
+from nsfrb.config import run_file,error_file
 def printlog(txt,output_file=run_file,end='\n'):
     if output_file != "":
         fout = open(output_file,"a")
@@ -25,14 +24,11 @@ def printlog(txt,output_file=run_file,end='\n'):
 #functions to send candidates to slack; adapted from https://api.slack.com/tutorials/tracks/uploading-files-python
 import logging, os
 from slack_sdk import WebClient
-error_file = cwd + "-logfiles/error_log.txt"
-cand_dir = os.environ['NSFRBDATA'] + "dsa110-nsfrb-candidates/"#cwd + "-candidates/"
-final_cand_dir = cand_dir + "final_cands/"
 
 #initialize Web API client
 client = WebClient(os.environ["SLACK_TOKEN_DSA"])
 candidates_channel_ID = "C07PVF82FQX"
-def send_candidate_slack(filename,filedir=final_cand_dir,error_file=error_file):
+def send_candidate_slack(filename,filedir,error_file=error_file):
     #upload file to bot
     try:
         client.files_upload_v2(channel=candidates_channel_ID,title=filename[:-4],file=filedir + filename,initial_comment=filename,)
@@ -44,7 +40,7 @@ def send_candidate_slack(filename,filedir=final_cand_dir,error_file=error_file):
 #slack alternative: pushover
 import requests
 import os
-def send_candidate_pushover(filename,filedir=final_cand_dir,error_file=error_file):
+def send_candidate_pushover(filename,filedir,error_file=error_file):
     try:
         r = requests.post("https://api.pushover.net/1/messages.json", data = {
             "token": os.environ["PSHOVR_TOKEN_DSA"],
@@ -66,7 +62,6 @@ from websockets.asyncio.client import connect
 from websockets.asyncio.server import serve
 import sys
 import json
-from nsfrb.config import frame_dir
 
 async def handler(websocket,filename):
     event = {"content":filename}
@@ -86,7 +81,7 @@ async def runserver(filename,port=9087):
         print("done")
     return
 
-def send_candidate_custom(filename,port=9087,filedir=final_cand_dir,error_file=error_file):
+def send_candidate_custom(filename,port=9087,error_file=error_file):
     try:
         asyncio.run(runserver(filename,port))
     except Exception as exc:
