+++ offline/offline_imager.py
@@ -174,7 +174,7 @@ def main(args):
         for g in range(len(filelabels)):
 
             #parameters from etcd
-            test, key_string, nant, nchan, npol, fobs, samples_per_frame, samples_per_frame_out, nint, nfreq_int, antenna_order, pt_dec, tsamp, fringestop, filelength_minutes, outrigger_delays, refmjd, subband = pu.parse_params(param_file=None,nsfrb=False)
+            test, key_string, nant, nchan, npol, fobs, samples_per_frame, samples_per_frame_out, nint, nfreq_int, antenna_order, pt_dec, tsamp, fringestop, filelength_minutes, outrigger_delays, refmjd, subband = pu.parse_params(param_file=None)
             #ff = 1.53-np.arange(8192)*0.25/8192
             #fobs = ff[1024:1024+int(len(corrs)*NUM_CHANNELS/2)]
             fobs = (np.reshape(freq_axis_fullres,(len(corrs)*args.nchans_per_node,int(NUM_CHANNELS/2/args.nchans_per_node))).mean(axis=1))*1e-3
